Rename payment_intent_ext
